Just the one is fine
