--- runelite-client/src/main/java/net/runelite/client/plugins/hideprayers/PVP/CombatPrayers.java
@@ -1,59 +1,0 @@
-/*
- * Copyright (c) 2018, Lotto <https://github.com/devLotto>
- * Copyright (c) 2018, Raqes <j.raqes@gmail.com>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *   list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *   this list of conditions and the following disclaimer in the documentation
- *   and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package net.runelite.client.plugins.hideprayers.PVP;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum CombatPrayers
-{
-	DISABLED("Disabled"),
-	PRAY1("1 prayer build"),
-	PRAY13("13 prayer build"),
-	PRAY16("16 prayer build"),
-	PRAY25("25 prayer build"),
-	PRAY31("31 prayer build"),
-	PRAY43("43 prayer build"),
-	PRAY44("44 prayer build"),
-	PRAY45("45 prayer build"),
-	PRAY52("52 prayer build"),
-	PRAY55("55 prayer build"),
-	PRAY60("60 prayer build"),
-	PRAY70("70 prayer build"),
-	PRAY74("74 prayer build"),
-	PRAY77("77 prayer build");
-
-	private String name;
-
-	@Override
-	public String toString()
-	{
-		return getName();
-	}
-}
