--- runelite-client/src/main/java/net/runelite/client/plugins/hideprayers/PVP/HidePrayersPVPConfig.java
@@ -1,60 +1,0 @@
-/*
- * Copyright (c) 2018, https://runelitepl.us
- * Copyright (c) 2019, Kyle <https://github.com/kyleeld>
- * Copyright (c) 2018, Raiever <https://github.com/Raieverr>
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- *    list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- *    this list of conditions and the following disclaimer in the documentation
- *    and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package net.runelite.client.plugins.hideprayers.PVP;
-
-import net.runelite.client.config.Config;
-import net.runelite.client.config.ConfigGroup;
-import net.runelite.client.config.ConfigItem;
-
-@ConfigGroup("hideprayersPVP")
-public interface HidePrayersPVPConfig extends Config
-{
-	@ConfigItem
-	(
-		position = 0,
-		keyName = "CombatPrayers",
-		name = "Combat Prayers",
-		description = "Shows prayers based on prayer build"
-	)
-	default CombatPrayers CombatPrayers()
-	{
-		return CombatPrayers.DISABLED;
-	}
-
-	@ConfigItem
-	(
-		position = 1,
-		keyName = "HideRapidHealRestore",
-		name = "Hide Rapid Heal and Rapid Restore",
-		description = "Hides the Rapid Heal and Rapid Restore prayers"
-	)
-	default boolean HideRapidHealRestore() 
-	{
-		return false;
-	}
-}
